// template/react-starter-kit/src/.generated/api/model/inquiriesControllerGetInquiriesSortItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type InquiriesControllerGetInquiriesSortItem = typeof InquiriesControllerGetInquiriesSortItem[keyof typeof InquiriesControllerGetInquiriesSortItem];


export const InquiriesControllerGetInquiriesSortItem = {
  createdAt: 'createdAt',
  updatedAt: 'updatedAt',
  status: 'status',
  title: 'title',
} as const;